+++ MusicPlayer2/TabDlg.cpp
@@ -97,7 +97,7 @@ void CTabDlg::ScrollWindowSimple(int step)
 		step = scrollinfo.nPos + step - scrollinfo.nMin;		//如果向上滚动一个距离后小于滚动条的最小位置了，则修正step的值，使窗口滚动到最上方
 		scrollinfo.nPos = scrollinfo.nMin;
 	}
-	if (scrollinfo.nPos + scrollinfo.nPage > scrollinfo.nMax)  //此处一定要注意加上滑块的长度，再作判断
+	if (scrollinfo.nPos + static_cast<int>(scrollinfo.nPage) > scrollinfo.nMax)  //此处一定要注意加上滑块的长度，再作判断
 	{
 		step -= (scrollinfo.nMax - (scrollinfo.nPos + scrollinfo.nPage));		//如果向上滚动一个距离后大于滚动条的最大位置了，则修正step的值，使窗口滚动到最下方
 		if (step > 0)
